requirements.txt
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0  # For PostgreSQL
pymysql>=1.0.0  # For MySQL
python-dotenv>=0.19.0
numpy>=1.19.5 
sentence-transformers>=3.0.0
torch>=2.0.0
flask>=2.0.0
flask-cors>=3.0.0
setuptools>=42.0.0
argparse>=1.4.0
webbrowser>=0.0.0
signal>=0.0.0
time>=0.0.0
traceback>=0.0.0
os>=0.0.0
sys>=0.0.0
subprocess>=0.0.0
atexit>=0.0.0
